+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Changed
 
+- Update some repo settings after creation as they are not taken correctly into account during creation by GitHub.
 - Include all organization settings when doing a rest call to update specific settings to ensure consistency. ([#14](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/14))
 - Added special handling for `web_commit_signoff_required`: if changed organization wide, it will implicitly change the same setting on repo level.
 - Removed `restrictsPushes` setting from branch protection rules as it is implicitly set based on setting `pushRestrictions`.